==> 기업가치/part2/04. 상관분석/02. 재무지표2/01. code/C4/ITS/2. corr_y_x_compare_period.ipynb <==
import pandas as pd
import numpy as np
import os
from scipy.cluster import hierarchy as hc
import matplotlib
import matplotlib.pyplot as plt
import warnings
import seaborn as sns
from scipy import stats
import math

%matplotlib inline
warnings.filterwarnings(action='ignore')

matplotlib.rcParams['font.family'] ='Malgun Gothic'
matplotlib.rcParams['axes.unicode_minus'] =False

dir_root = '../../../../../'
current_dir = os.getcwd()
dir_master = os.path.join(dir_root, '01. Data','00. master_data')
dir_processed = os.path.join(dir_root, '01. Data','02. 재무지표2', '01. preprocessed')
dir_result = os.path.join(dir_root, '04. 상관분석','02. 재무지표2','02. result')

# Accenture

## Load data

company = 'Accenture'
y = '유동비율(비율)'

dir_company = os.path.join(dir_result, company)
dir_company_y = os.path.join(dir_company, y)

result = pd.read_csv(os.path.join(dir_company_y, 'result.csv'))
result_sub = pd.read_csv(os.path.join(dir_company_y, 'result_sub.csv'))

## Compare : total ~ sub

### over 0.7

result_total_7 = pd.merge(result[np.abs(result['corr'])>0.7], result_sub[np.abs(result_sub['corr'])>0.7], 
                          on=['company', 'variable', 'group', 'importance'], how='outer',
                         suffixes = ("_total", "_sub"))
result_total_7 = result_total_7.sort_values(['corr_total'], ascending=False)

result_total_7 = result_total_7[['company', 'variable', 'group', 'importance', 'corr_total', 'corr_sub']]

result_total_7.to_csv(os.path.join(dir_company_y, 'compare_period_corr_07.csv'), index=False, encoding='utf-8-sig')

### over 0.5

result_total_5 = pd.merge(result[np.abs(result['corr'])>0.5], result_sub[np.abs(result_sub['corr'])>0.5], 
                          on=['company', 'variable', 'group', 'importance'], how='outer',
                         suffixes = ("_total", "_sub"))

result_total_5 = result_total_5.sort_values(['corr_total'], ascending=False)

result_total_5 = result_total_5[['company', 'variable', 'group', 'importance', 'corr_total', 'corr_sub']]

result_total_5.to_csv(os.path.join(dir_company_y, 'compare_period_corr_05.csv'), index=False, encoding='utf-8-sig')

### Save

writer = pd.ExcelWriter(os.path.join(dir_company_y, y+'.xlsx'), engine='xlsxwriter')
pd.concat([result, result_sub], axis=1).to_excel(writer, sheet_name='Correlation', index=False)
result_total_7.to_excel(writer, sheet_name='Correlation > 0.7', index=False)
result_total_5.to_excel(writer, sheet_name='Correlation > 0.5', index=False)
writer.save()

# Infosys

## Load data

company = 'Infosys'
y = '유동비율(비율)'

dir_company = os.path.join(dir_result, company)
dir_company_y = os.path.join(dir_company, y)

result = pd.read_csv(os.path.join(dir_company_y, 'result.csv'))
result_sub = pd.read_csv(os.path.join(dir_company_y, 'result_sub.csv'))

## Compare : total ~ sub

### over 0.7

result_total_7 = pd.merge(result[np.abs(result['corr'])>0.7], result_sub[np.abs(result_sub['corr'])>0.7], 
                          on=['company', 'variable', 'group', 'importance'], how='outer',
                         suffixes = ("_total", "_sub"))
result_total_7 = result_total_7.sort_values(['corr_total'], ascending=False)

result_total_7 = result_total_7[['company', 'variable', 'group', 'importance', 'corr_total', 'corr_sub']]

result_total_7.to_csv(os.path.join(dir_company_y, 'compare_period_corr_07.csv'), index=False, encoding='utf-8-sig')

### over 0.5

result_total_5 = pd.merge(result[np.abs(result['corr'])>0.5], result_sub[np.abs(result_sub['corr'])>0.5], 
                          on=['company', 'variable', 'group', 'importance'], how='outer',
                         suffixes = ("_total", "_sub"))

result_total_5 = result_total_5.sort_values(['corr_total'], ascending=False)

result_total_5 = result_total_5[['company', 'variable', 'group', 'importance', 'corr_total', 'corr_sub']]

result_total_5.to_csv(os.path.join(dir_company_y, 'compare_period_corr_05.csv'), index=False, encoding='utf-8-sig')

### Save

writer = pd.ExcelWriter(os.path.join(dir_company_y, y+'.xlsx'), engine='xlsxwriter')
pd.concat([result, result_sub], axis=1).to_excel(writer, sheet_name='Correlation', index=False)
result_total_7.to_excel(writer, sheet_name='Correlation > 0.7', index=False)
result_total_5.to_excel(writer, sheet_name='Correlation > 0.5', index=False)
writer.save()

# TCS

## Load data

company = 'TCS'
y = '유동비율(비율)'

dir_company = os.path.join(dir_result, company)
dir_company_y = os.path.join(dir_company, y)

result = pd.read_csv(os.path.join(dir_company_y, 'result.csv'))
result_sub = pd.read_csv(os.path.join(dir_company_y, 'result_sub.csv'))

## Compare : total ~ sub

### over 0.7

result_total_7 = pd.merge(result[np.abs(result['corr'])>0.7], result_sub[np.abs(result_sub['corr'])>0.7], 
                          on=['company', 'variable', 'group', 'importance'], how='outer',
                         suffixes = ("_total", "_sub"))
result_total_7 = result_total_7.sort_values(['corr_total'], ascending=False)

result_total_7 = result_total_7[['company', 'variable', 'group', 'importance', 'corr_total', 'corr_sub']]

result_total_7.to_csv(os.path.join(dir_company_y, 'compare_period_corr_07.csv'), index=False, encoding='utf-8-sig')

### over 0.5

result_total_5 = pd.merge(result[np.abs(result['corr'])>0.5], result_sub[np.abs(result_sub['corr'])>0.5], 
                          on=['company', 'variable', 'group', 'importance'], how='outer',
                         suffixes = ("_total", "_sub"))

result_total_5 = result_total_5.sort_values(['corr_total'], ascending=False)

result_total_5 = result_total_5[['company', 'variable', 'group', 'importance', 'corr_total', 'corr_sub']]

result_total_5.to_csv(os.path.join(dir_company_y, 'compare_period_corr_05.csv'), index=False, encoding='utf-8-sig')

### Save

writer = pd.ExcelWriter(os.path.join(dir_company_y, y+'.xlsx'), engine='xlsxwriter')
pd.concat([result, result_sub], axis=1).to_excel(writer, sheet_name='Correlation', index=False)
result_total_7.to_excel(writer, sheet_name='Correlation > 0.7', index=False)
result_total_5.to_excel(writer, sheet_name='Correlation > 0.5', index=False)
writer.save()

# Cognizant

## Load data

company = 'Cognizant'
y = '유동비율(비율)'

dir_company = os.path.join(dir_result, company)
dir_company_y = os.path.join(dir_company, y)

result = pd.read_csv(os.path.join(dir_company_y, 'result.csv'))
result_sub = pd.read_csv(os.path.join(dir_company_y, 'result_sub.csv'))

## Compare : total ~ sub

### over 0.7

result_total_7 = pd.merge(result[np.abs(result['corr'])>0.7], result_sub[np.abs(result_sub['corr'])>0.7], 
                          on=['company', 'variable', 'group', 'importance'], how='outer',
                         suffixes = ("_total", "_sub"))
result_total_7 = result_total_7.sort_values(['corr_total'], ascending=False)

result_total_7 = result_total_7[['company', 'variable', 'group', 'importance', 'corr_total', 'corr_sub']]

result_total_7.to_csv(os.path.join(dir_company_y, 'compare_period_corr_07.csv'), index=False, encoding='utf-8-sig')

### over 0.5

result_total_5 = pd.merge(result[np.abs(result['corr'])>0.5], result_sub[np.abs(result_sub['corr'])>0.5], 
                          on=['company', 'variable', 'group', 'importance'], how='outer',
                         suffixes = ("_total", "_sub"))

result_total_5 = result_total_5.sort_values(['corr_total'], ascending=False)

result_total_5 = result_total_5[['company', 'variable', 'group', 'importance', 'corr_total', 'corr_sub']]

result_total_5.to_csv(os.path.join(dir_company_y, 'compare_period_corr_05.csv'), index=False, encoding='utf-8-sig')

### Save

writer = pd.ExcelWriter(os.path.join(dir_company_y, y+'.xlsx'), engine='xlsxwriter')
pd.concat([result, result_sub], axis=1).to_excel(writer, sheet_name='Correlation', index=False)
result_total_7.to_excel(writer, sheet_name='Correlation > 0.7', index=False)
result_total_5.to_excel(writer, sheet_name='Correlation > 0.5', index=False)
writer.save()
